requirements.txt
# API Integration
requests>=2.28.0

# Data Handling and Parsing
pdfminer.six>=20221105
PyMuPDF>=1.21.0
beautifulsoup4>=4.11.1
python-docx>=0.8.11


# Natural Language Processing
spacy>=3.5.0
nltk>=3.8.1
thefuzz>=0.19.0
python-Levenshtein>=0.21.0

# Document Comparison & Similarity
scikit-learn>=1.2.0
sentence-transformers>=2.2.2

# Output & Reporting
pandas>=1.5.0

# Configuration
pyyaml>=6.0

# Testing
pytest>=7.3.1

# Optional visualization/UI (uncomment if needed)
streamlit>=1.21.0

# Install all requirements
pip install -r requirements.txt
python -m spacy download en_core_web_sm

# Install your package in development mode
pip install -e 
